fillMatrix func fixed

+++ muraviev.alexander/P3/main.cpp
@@ -113,7 +113,7 @@ namespace muraviev
 
   void fillMatrix(std::ifstream& fin, int * matrix, size_t rows, size_t columns)
   {
-    for (int i = 0; i < rows * columns; ++i) {
+    for (size_t i = 0; i < rows * columns; ++i) {
       if (!(fin >> matrix[i])) {
         throw std::logic_error("func fillMatrix failed with filling elements.");
       }
